# packages/api-a-hono/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API A (Profile Service) — Typed SDK Demo",
    "version": "0.1.0",
    "description": "Zod-first banking profile API → OpenAPI → generated SDK."
  },
  "servers": [
    {
      "url": "http://localhost:4011",
      "description": "Local dev"
    }
  ],
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [true]
          }
        },
        "required": ["ok"]
      },
      "CustomerStatus": {
        "type": "string",
        "enum": ["active", "inactive", "frozen"]
      },
      "Province": {
        "type": "string",
        "enum": [
          "AB",
          "BC",
          "MB",
          "NB",
          "NL",
          "NS",
          "NT",
          "NU",
          "ON",
          "PE",
          "QC",
          "SK",
          "YT"
        ]
      },
      "Address": {
        "type": "object",
        "properties": {
          "line1": {
            "type": "string",
            "minLength": 1
          },
          "line2": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "minLength": 1
          },
          "province": {
            "type": "string",
            "enum": [
              "AB",
              "BC",
              "MB",
              "NB",
              "NL",
              "NS",
              "NT",
              "NU",
              "ON",
              "PE",
              "QC",
              "SK",
              "YT"
            ]
          },
          "postalCode": {
            "type": "string",
            "minLength": 3
          },
          "country": {
            "type": "string",
            "enum": ["CA"]
          }
        },
        "required": ["line1", "city", "province", "postalCode", "country"]
      },
      "CustomerProfile": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "minLength": 1
          },
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "minLength": 7
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "frozen"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "object",
            "properties": {
              "line1": {
                "type": "string",
                "minLength": 1
              },
              "line2": {
                "type": "string"
              },
              "city": {
                "type": "string",
                "minLength": 1
              },
              "province": {
                "type": "string",
                "enum": [
                  "AB",
                  "BC",
                  "MB",
                  "NB",
                  "NL",
                  "NS",
                  "NT",
                  "NU",
                  "ON",
                  "PE",
                  "QC",
                  "SK",
                  "YT"
                ]
              },
              "postalCode": {
                "type": "string",
                "minLength": 3
              },
              "country": {
                "type": "string",
                "enum": ["CA"]
              }
            },
            "required": ["line1", "city", "province", "postalCode", "country"]
          }
        },
        "required": [
          "customerId",
          "fullName",
          "email",
          "status",
          "createdAt",
          "address"
        ]
      },
      "CustomerSearchRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "frozen"]
          },
          "province": {
            "type": "string",
            "enum": [
              "AB",
              "BC",
              "MB",
              "NB",
              "NL",
              "NS",
              "NT",
              "NU",
              "ON",
              "PE",
              "QC",
              "SK",
              "YT"
            ]
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "default": 10
          }
        }
      },
      "CustomerSearchResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "customerId": {
                  "type": "string",
                  "minLength": 1
                },
                "fullName": {
                  "type": "string",
                  "minLength": 1
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "phone": {
                  "type": "string",
                  "minLength": 7
                },
                "status": {
                  "type": "string",
                  "enum": ["active", "inactive", "frozen"]
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "address": {
                  "type": "object",
                  "properties": {
                    "line1": {
                      "type": "string",
                      "minLength": 1
                    },
                    "line2": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string",
                      "minLength": 1
                    },
                    "province": {
                      "type": "string",
                      "enum": [
                        "AB",
                        "BC",
                        "MB",
                        "NB",
                        "NL",
                        "NS",
                        "NT",
                        "NU",
                        "ON",
                        "PE",
                        "QC",
                        "SK",
                        "YT"
                      ]
                    },
                    "postalCode": {
                      "type": "string",
                      "minLength": 3
                    },
                    "country": {
                      "type": "string",
                      "enum": ["CA"]
                    }
                  },
                  "required": [
                    "line1",
                    "city",
                    "province",
                    "postalCode",
                    "country"
                  ]
                }
              },
              "required": [
                "customerId",
                "fullName",
                "email",
                "status",
                "createdAt",
                "address"
              ]
            }
          }
        },
        "required": ["results"]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [true]
                    }
                  },
                  "required": ["ok"]
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customerId}": {
      "get": {
        "summary": "Get customer profile by customerId",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "customerId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customerId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "fullName": {
                      "type": "string",
                      "minLength": 1
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "phone": {
                      "type": "string",
                      "minLength": 7
                    },
                    "status": {
                      "type": "string",
                      "enum": ["active", "inactive", "frozen"]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "address": {
                      "type": "object",
                      "properties": {
                        "line1": {
                          "type": "string",
                          "minLength": 1
                        },
                        "line2": {
                          "type": "string"
                        },
                        "city": {
                          "type": "string",
                          "minLength": 1
                        },
                        "province": {
                          "type": "string",
                          "enum": [
                            "AB",
                            "BC",
                            "MB",
                            "NB",
                            "NL",
                            "NS",
                            "NT",
                            "NU",
                            "ON",
                            "PE",
                            "QC",
                            "SK",
                            "YT"
                          ]
                        },
                        "postalCode": {
                          "type": "string",
                          "minLength": 3
                        },
                        "country": {
                          "type": "string",
                          "enum": ["CA"]
                        }
                      },
                      "required": [
                        "line1",
                        "city",
                        "province",
                        "postalCode",
                        "country"
                      ]
                    }
                  },
                  "required": [
                    "customerId",
                    "fullName",
                    "email",
                    "status",
                    "createdAt",
                    "address"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/customers/search": {
      "post": {
        "summary": "Search customers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "inactive", "frozen"]
                  },
                  "province": {
                    "type": "string",
                    "enum": [
                      "AB",
                      "BC",
                      "MB",
                      "NB",
                      "NL",
                      "NS",
                      "NT",
                      "NU",
                      "ON",
                      "PE",
                      "QC",
                      "SK",
                      "YT"
                    ]
                  },
                  "limit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50,
                    "default": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "customerId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "fullName": {
                            "type": "string",
                            "minLength": 1
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "phone": {
                            "type": "string",
                            "minLength": 7
                          },
                          "status": {
                            "type": "string",
                            "enum": ["active", "inactive", "frozen"]
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "address": {
                            "type": "object",
                            "properties": {
                              "line1": {
                                "type": "string",
                                "minLength": 1
                              },
                              "line2": {
                                "type": "string"
                              },
                              "city": {
                                "type": "string",
                                "minLength": 1
                              },
                              "province": {
                                "type": "string",
                                "enum": [
                                  "AB",
                                  "BC",
                                  "MB",
                                  "NB",
                                  "NL",
                                  "NS",
                                  "NT",
                                  "NU",
                                  "ON",
                                  "PE",
                                  "QC",
                                  "SK",
                                  "YT"
                                ]
                              },
                              "postalCode": {
                                "type": "string",
                                "minLength": 3
                              },
                              "country": {
                                "type": "string",
                                "enum": ["CA"]
                              }
                            },
                            "required": [
                              "line1",
                              "city",
                              "province",
                              "postalCode",
                              "country"
                            ]
                          }
                        },
                        "required": [
                          "customerId",
                          "fullName",
                          "email",
                          "status",
                          "createdAt",
                          "address"
                        ]
                      }
                    }
                  },
                  "required": ["results"]
                }
              }
            }
          }
        }
      }
    }
  }
}